refactor: moved files related to ext only
